refactor(session-store): extract canonical graph-builders (file #4)
Lift the immutable SessionStateGraph copy helpers (copy-keys + graphWith{TranscriptSnapshot,
Lifecycle,Capabilities,Patches}) + the session-export content-error types/factory into
session-graph-builders.ts; session-store re-exports the error types. Pure graph transforms.
check 0, store vitest parity (503 pass).
